# hw2/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: API for marketplace with products, orders, and promo codes
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get list of products with pagination and filtering
      description: Available to all authenticated users
      parameters:
        - name: page
          in: query
          description: Page number (starting from 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by product status
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Filter by category (exact match)
          schema:
            type: string
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Products
      summary: Create a new product
      description: Available to SELLER (own products) and ADMIN (any products)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Available to all authenticated users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Products
      summary: Update product
      description: SELLER can update only own products, ADMIN can update any
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Products
      summary: Soft delete product (set status to ARCHIVED)
      description: SELLER can delete only own products, ADMIN can delete any
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product archived successfully
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Order endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
        - category
        - status
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Product description
        price:
          type: number
          format: decimal
          minimum: 0.01
          exclusiveMinimum: true
          description: Product price (must be greater than 0)
        stock:
          type: integer
          minimum: 0
          description: Stock quantity
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Product category
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
          exclusiveMinimum: true
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ARCHIVED
      description: Product status
    ProductListResponse:
      type: object
      required:
        - items
        - total_elements
        - page
        - size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        total_elements:
          type: integer
          description: Total number of products matching the filter
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Page size

    # Order schemas
...
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          enum:
            - PRODUCT_NOT_FOUND
            - PRODUCT_INACTIVE
            - ORDER_NOT_FOUND
            - ORDER_LIMIT_EXCEEDED
            - ORDER_HAS_ACTIVE
            - INVALID_STATE_TRANSITION
            - INSUFFICIENT_STOCK
            - PROMO_CODE_INVALID
            - PROMO_CODE_MIN_AMOUNT
            - ORDER_OWNERSHIP_VIOLATION
            - VALIDATION_ERROR
            - ACCESS_DENIED
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - REFRESH_TOKEN_INVALID
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error details
          additionalProperties: true